server/pyproject.toml
[project]
name = "voice-live-agent"
version = "0.1.0"
description = "Real-time voice agent with Azure Voice Live and ACS/Web streaming clients"
dependencies = [
    "quart",
    "websockets",
    "httpx",
    "python-dotenv",
    "quart-cors",
    "azure-core",
    "azure-communication-identity",
    "aiohttp",
    "azure-eventgrid",
    "azure-identity",
    "azure-communication-callautomation",
    "openai[realtime]",
    "pydantic",
    "pytest",
    "pytest-asyncio",
    "jsonschema"
]
requires-python = ">=3.9"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
